backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
PyJWT==2.9.0
bcrypt==4.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12